# Add lru_cache on parse for observe to speed up instantiation time