@aaronayres35 2020-11-09
see https://github.com/enthought/traits/pull/1343#issuecomment-724215711

Im copying this comment here to continue the discussion on this PR.  I realize the original question I wrote wasn't great.  The `functools.lru_cache` docs say "Decorator to wrap a function with a memoizing callable that saves up to the `maxsize` most recent calls.".  So the cache will hold 128 calls to the function, i.e. in response to:

>  Perhaps we could try to work out how many ObserverExpression the cache can hold

I believe it can hold `maxsize` `ObserverExpression`'s, or more specifically it can hold the returned result of 128 different `text` inputs to the `parse` function.  

I'm still unsure what a realistic number for `maxsize` would be in practice here