@kitchoi 2020-11-10
Thanks. For some reason I thought the limit unit is in memory space, sorry about that. This makes sense.
Then we should probably estimate how much memory a full cache is going to take? Even if it should result in no action (as in, it does not take a good deal of memory), it would be good to know for future reference. And if it should take up a lot of space, we might want to look into optimizing memory usage.