@aaronayres35 2020-11-10
> Then we should probably estimate how much memory a full cache is going to take? Even if it should result in no action (as in, it does not take a good deal of memory), it would be good to know for future reference. And if it should take up a lot of space, we might want to look into optimizing memory usage.

@kitchoi an instance of `ObserverExpression` is 56 bytes (according to `sys.getsizeof`, note from the docs: "Only the memory consumption directly attributed to the object is accounted for, not the memory consumption of objects it refers to.")
```
$ python
>>> import sys
>>> from traits.observation.expression import ObserverExpression
>>> sys.getsizeof(ObserverExpression())
56
```

likewise a sanity check of calling `sys.getsizeof` on what is returned by the `parse` method when running the benchmarking script with `n=1` also gives 56 bytes.

As far as the full memory usage of the cache I'm not certain, as I don't know _exactly_ what all the cache is storing (it has to at least also store the inputs to the function, I believe it keeps a dictionary of the form {input: output}).  Based on this stack exchange discussion https://stackoverflow.com/questions/23477284/memory-aware-lru-caching-in-python it seems we could get a better estimate using `psutil`, but I am not sure if it is worth it to investigate.  Another quick check I did was the following:

```
$ python
>>> import sys
>>> from traits.observation.expression import ObserverExpression
>>> sys.getsizeof({'some.string': ObserverExpression()})
240
```